Refactored Room and Person classes.  Updated Slack API calls and logic.
