+++ _slack/person.py
@@ -1,8 +1,7 @@
 import logging
 
-from slack_sdk.web import WebClient
-
 from errbot.backends.base import Person, RoomDoesNotExistError
+from slack_sdk.web import WebClient
 
 log = logging.getLogger(__name__)
 
@@ -147,3 +146,4 @@ class SlackPerson(Person):
 
     def __hash__(self):
         return self.userid.__hash__()
+
