+++ slackv3.py
@@ -2,9 +2,9 @@ import copyreg
 import json
 import logging
 import pprint
-import threading
 import re
 import sys
+import threading
 from functools import lru_cache
 from typing import BinaryIO
 
@@ -35,10 +35,10 @@ log = logging.getLogger(__name__)
 try:
     from slack_sdk.errors import BotUserAccessError, SlackApiError
     from slack_sdk.rtm.v2 import RTMClient
-    from slack_sdk.web import WebClient
     from slack_sdk.socket_mode import SocketModeClient
-    from slack_sdk.socket_mode.response import SocketModeResponse
     from slack_sdk.socket_mode.request import SocketModeRequest
+    from slack_sdk.socket_mode.response import SocketModeResponse
+    from slack_sdk.web import WebClient
     from slackeventsapi import SlackEventAdapter
 
 except ImportError:
@@ -49,9 +49,10 @@ except ImportError:
     )
     sys.exit(1)
 
+from _slack.lib import USER_IS_BOT_HELPTEXT, SlackAPIResponseError
 from _slack.markdown import slack_markdown_converter
 from _slack.person import SlackPerson
-
+from _slack.room import SlackRoom, SlackRoomBot, SlackRoomOccupant
 
 # The Slack client automatically turns a channel name into a clickable
 # link if you prefix it with a #. Other clients receive this link as a
@@ -76,106 +77,6 @@ COLORS = {
     "white": "#FFFFFF",
     "cyan": "#00FFFF",
 }  # Slack doesn't know its colors
-
-
-class SlackAPIResponseError(RuntimeError):
-    """Slack API returned a non-OK response"""
-
-    def __init__(self, *args, error="", **kwargs):
-        """
-        :param error:
-            The 'error' key from the API response data
-        """
-        self.error = error
-        super().__init__(*args, **kwargs)
-
-
-class SlackRoomOccupant(RoomOccupant, SlackPerson):
-    """
-    This class represents a person inside a MUC.
-    """
-
-    def __init__(self, webclient: WebClient, userid, channelid, bot):
-        super().__init__(webclient, userid, channelid)
-        self._room = SlackRoom(webclient=webclient, channelid=channelid, bot=bot)
-
-    @property
-    def room(self):
-        return self._room
-
-    def __unicode__(self):
-        return f"#{self._room.name}/{self.username}"
-
-    def __str__(self):
-        return self.__unicode__()
-
-    def __eq__(self, other):
-        if not isinstance(other, SlackRoomOccupant):
-            log.warning(
-                "tried to compare a SlackRoomOccupant with a SlackPerson %s vs %s",
-                self,
-                other,
-            )
-            return False
-        return other.room.id == self.room.id and other.userid == self.userid
-
-
-class SlackBot(SlackPerson):
-    """
-    This class describes a bot on Slack's network.
-    """
-
-    def __init__(self, webclient: WebClient, bot_id, bot_username):
-        self._bot_id = bot_id
-        self._bot_username = bot_username
-        super().__init__(webclient, userid=bot_id)
-
-    @property
-    def username(self):
-        return self._bot_username
-
-    # Beware of gotcha. Without this, nick would point to username of SlackPerson.
-    nick = username
-
-    @property
-    def aclattr(self):
-        # Make ACLs match against integration ID rather than human-readable
-        # nicknames to avoid webhooks impersonating other people.
-        return f"<{self._bot_id}>"
-
-    @property
-    def fullname(self):
-        return None
-
-
-class SlackRoomBot(RoomOccupant, SlackBot):
-    """
-    This class represents a bot inside a MUC.
-    """
-
-    def __init__(self, sc, bot_id, bot_username, channelid, bot):
-        super().__init__(sc, bot_id, bot_username)
-        self._room = SlackRoom(webclient=sc, channelid=channelid, bot=bot)
-
-    @property
-    def room(self):
-        return self._room
-
-    def __unicode__(self):
-        return f"#{self._room.name}/{self.username}"
-
-    def __str__(self):
-        return self.__unicode__()
-
-    def __eq__(self, other):
-        if not isinstance(other, SlackRoomOccupant):
-            log.warning(
-                "tried to compare a SlackRoomBotOccupant with a SlackPerson %s vs %s",
-                self,
-                other,
-            )
-            return False
-        return other.room.id == self.room.id and other.userid == self.userid
 
 
 class SlackBackend(ErrBot):
@@ -674,22 +575,15 @@ class SlackBackend(ErrBot):
 
     def channelid_to_channelname(self, id_: str):
         """Convert a Slack channel ID to its channel name"""
-        channel = self.slack_web.conversations_info(channel=id_)["channel"]
-        if channel is None:
-            raise RoomDoesNotExistError(f"No channel with ID {id_} exists.")
-        return channel["name"]
+        log.warning(f"get channel name from {id_}")
+        room = SlackRoom(self._webclient, channelid=id_, bot=self)
+        return room.channelname
 
     def channelname_to_channelid(self, name: str):
         """Convert a Slack channel name to its channel ID"""
-        name = name.lstrip("#")
-        channel = [
-            channel
-            for channel in self.slack_web.conversations_list()["channels"]
-            if channel["name"] == name
-        ]
-        if not channel:
-            raise RoomDoesNotExistError(f"No channel named {name} exists")
-        return channel[0]["id"]
+        log.warning(f"get channel id from {name}")
+        room = SlackRoom(self._webclient, name=name, bot=self)
+        return room.id
 
     def channels(
         self,
@@ -1121,7 +1015,7 @@ class SlackBackend(ErrBot):
 
             ts = self._ts_for_message(msg)
 
-            self.api_call(
+            self.slackweb.api_call(
                 method,
                 data={"channel": to_channel_id, "timestamp": ts, "name": reaction},
             )
@@ -1229,217 +1123,3 @@ class SlackBackend(ErrBot):
                 text = text.replace(word, f"#{identifier.channelid}")
 
         return text, mentioned
-
-
-class SlackRoom(Room):
-    def __init__(self, webclient=None, name=None, channelid=None, bot=None):
-        if channelid is not None and name is not None:
-            raise ValueError("channelid and name are mutually exclusive")
-
-        if name is not None:
-            if name.startswith("#"):
-                self._name = name[1:]
-            else:
-                self._name = name
-        else:
-            self._name = bot.channelid_to_channelname(channelid)
-
-        self._id = channelid
-        self._bot = bot
-        self.slack_web = webclient
-
-    def __str__(self):
-        return f"#{self.name}"
-
-    @property
-    def channelname(self):
-        return self._name
-
-    @property
-    def _channel(self):
-        """
-        The channel object exposed by SlackClient
-        """
-        log.warning("Resolving channel '%s' by iterating all channels", self.name)
-        channel_id = None
-        try:
-            cursor = None
-            while True:
-                conversations_list = self.slack_web.conversations_list(cursor=cursor, limit=1000)
-                for channel in conversations_list["channels"]:
-                    if channel["name"] == self.name:
-                        channel_id = channel["id"]
-                        raise StopIteration
-                cursor = conversations_list["response_metadata"].get("next_cursor", None)
-                if not cursor:
-                    raise RoomDoesNotExistError(f"Cannot find channel {self.name}.")
-        except StopIteration:
-            pass
-        log.warning("Channel '%s' resolved to channel id '%s'", self.name, channel_id)
-        return channel_id
-
-    @property
-    def _channel_info(self):
-        """
-        Channel info as returned by the Slack API.
-
-        See also:
-          * https://api.slack.com/methods/channels.list
-          * https://api.slack.com/methods/groups.list
-        """
-        return self._bot.slack_web.conversations_info(channel=self.id)["channel"]
-
-    @property
-    def private(self):
-        """Return True if the room is a private group"""
-        return self._channel_info["is_private"]
-
-    @property
-    def id(self):
-        """Return the ID of this room"""
-        if self._id is None:
-            self._id = self._channel
-        return self._id
-
-    @property
-    def name(self):
-        """Return the name of this room"""
-        return self._name
-
-    def join(self, username=None, password=None):
-        log.info("Joining channel '{}'".format(self.name))
-        join_failure = True
-        try:
-            self._bot.slack_web.conversations_join(channel=self.id)
-            join_failure = False
-        except SlackApiError as e:
-            log.error(f"Unable to join '{self.name}'. Slack API Error {str(e)}")
-        except BotUserAccessError as e:
-            log.error(f"OAuthv1 bot token not allowed to join channels. '{self.name}'.")
-
-        if join_failure:
-            raise RoomError(f"Unable to join channel. {USER_IS_BOT_HELPTEXT}")
-
-    def leave(self, reason=None):
-        try:
-            log.info("Leaving conversation %s (%s)", self, self.id)
-            self._bot.slack_web.conversations_leave(channel=self.id)
-        except SlackAPIResponseError as e:
-            if e.error == "user_is_bot":
-                raise RoomError(f"Unable to leave channel. {USER_IS_BOT_HELPTEXT}")
-            else:
-                raise RoomError(e)
-        self._id = None
-
-    def create(self, private=False):
-        try:
-            if private:
-                log.info("Creating private conversation %s.", self)
-                self._bot.slack_web.conversations_create(
-                    name=self.name, is_private=True
-                )
-            else:
-                log.info("Creating conversation %s.", self)
-                self._bot.slack_web.conversations_create(name=self.name)
-        except SlackAPIResponseError as e:
-            if e.error == "user_is_bot":
-                raise RoomError(f"Unable to create channel. {USER_IS_BOT_HELPTEXT}")
-            else:
-                raise RoomError(e)
-
-    def destroy(self):
-        try:
-            log.info("Archiving conversation %s (%s)", self, self.id)
-            self._bot.slack_web.conversations_archive(self.id)
-        except SlackAPIResponseError as e:
-            if e.error == "user_is_bot":
-                raise RoomError(f"Unable to archive channel. {USER_IS_BOT_HELPTEXT}")
-            else:
-                raise RoomError(e)
-        self._id = None
-
-    @property
-    def exists(self):
-        channels = self._bot.channels(joined_only=False, exclude_archived=False)
-        return len([c for c in channels if c["name"] == self.name]) > 0
-
-    @property
-    def joined(self):
-        channels = self._bot.channels(joined_only=True)
-        return len([c for c in channels if c["name"] == self.name]) > 0
-
-    @property
-    def topic(self):
-        if self._channel_info["topic"]["value"] == "":
-            return None
-        else:
-            return self._channel_info["topic"]["value"]
-
-    @topic.setter
-    def topic(self, topic):
-        if self.private:
-            log.info("Setting topic of %s (%s) to %s.", self, self.id, topic)
-            self._bot.api_call(
-                "groups.setTopic", data={"channel": self.id, "topic": topic}
-            )
-        else:
-            log.info("Setting topic of %s (%s) to %s.", self, self.id, topic)
-            self._bot.api_call(
-                "channels.setTopic", data={"channel": self.id, "topic": topic}
-            )
-
-    @property
-    def purpose(self):
-        if self._channel_info["purpose"]["value"] == "":
-            return None
-        else:
-            return self._channel_info["purpose"]["value"]
-
-    @purpose.setter
-    def purpose(self, purpose):
-        if self.private:
-            log.info("Setting purpose of %s (%s) to %s.", self, self.id, purpose)
-            self._bot.api_call(
-                "groups.setPurpose", data={"channel": self.id, "purpose": purpose}
-            )
-        else:
-            log.info("Setting purpose of %s (%s) to %s.", str(self), self.id, purpose)
-            self._bot.api_call(
-                "channels.setPurpose", data={"channel": self.id, "purpose": purpose}
-            )
-
-    @property
-    def occupants(self):
-        members = self._channel_info["members"]
-        return [
-            SlackRoomOccupant(self.slack_web, m, self.id, self._bot) for m in members
-        ]
-
-    def invite(self, *args):
-        users = {
-            user["name"]: user["id"]
-            for user in self._bot.api_call("users.list")["members"]
-        }
-        for user in args:
-            if user not in users:
-                raise UserDoesNotExistError(f'User "{user}" not found.')
-            log.info("Inviting %s into %s (%s)", user, self, self.id)
-            method = "groups.invite" if self.private else "channels.invite"
-            response = self._bot.api_call(
-                method,
-                data={"channel": self.id, "user": users[user]},
-                raise_errors=False,
-            )
-
-            if not response["ok"]:
-                if response["error"] == "user_is_bot":
-                    raise RoomError(f"Unable to invite people. {USER_IS_BOT_HELPTEXT}")
-                elif response["error"] != "already_in_channel":
-                    raise SlackAPIResponseError(
-                        error=f'Slack API call to {method} failed: {response["error"]}.'
-                    )
-
-    def __eq__(self, other):
-        if not isinstance(other, SlackRoom):
-            return False
-        return self.id == other.id
